Fix users foreign key types for daily logs

+++ db/migrations/20260323_add_daily_logs.sql
@@ -1,10 +1,10 @@
 CREATE TABLE IF NOT EXISTS daily_logs (
   id BIGINT PRIMARY KEY AUTO_INCREMENT,
-  user_id BIGINT NOT NULL,
+  user_id INT NOT NULL,
   entry_date DATE NOT NULL,
   submitted TINYINT(1) NOT NULL DEFAULT 0,
-  created_by BIGINT NULL,
-  updated_by BIGINT NULL,
+  created_by INT NULL,
+  updated_by INT NULL,
   created_at DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP,
   updated_at DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
   UNIQUE KEY uq_daily_logs_user_date (user_id, entry_date),
